+++ scripts/test_python.sh
@@ -33,8 +33,11 @@ test_crate=${1:-""}
 if [ -z "$test_crate" ]; then
     test_python_interface findex
     test_python_interface cover_crypt
+    test_python_interface fpe
 elif [ "$test_crate" = "findex" ]; then
     test_python_interface findex
 elif [ "$test_crate" = "cover_crypt" ]; then
     test_python_interface cover_crypt
+elif [ "$test_crate" = "fpe" ]; then
+    test_python_interface fpe
 fi